+++ document-service/src/main/java/com/buaa01/illumineer_backend/service/paper/PaperStatsService.java
@@ -4,5 +4,5 @@ import com.buaa01.illumineer_backend.entity.CustomResponse;
 
 public interface PaperStatsService {
 
-    CustomResponse updateStats(Integer pid, Integer stats);
+    CustomResponse updateStats(Long pid, Integer stats);
 }
